+++ hosts/sigma/configuration.nix
@@ -25,8 +25,6 @@
 
   services.fprintd.enable = true;
 
-  services.xserver.displayManager.defaultSession = lib.mkForce "sway";
-
   # Check that this can be bumped before changing it
   system.stateVersion = "22.05";
 }