+++ modules/vm.nix
@@ -26,8 +26,6 @@
     programs.sway.extraSessionCommands = ''
       export WLR_NO_HARDWARE_CURSORS=1
     '';
-
-    services.xserver.displayManager.defaultSession = mkIf config.programs.sway.enable (mkVMOverride "sway");
   };
 
   # WORKAROUND: { osConfig ? { }, ... }: fails when using `home-manager build`